Display first word of multiline code blocks
When receiving a direct message with text on the first line of a
multi-line code block, libpurple-based clients do not show the
characters up to the first space or newline character.  For example,
sending the text from example 2 of [Markdown Text
101](https://support.discordapp.com/hc/en-us/articles/210298617-Markdown-Text-101-Chat-Formatting-Bold-Italic-Underline-):

    ```haikus written here
    really shouldn't cause conflict
    markdown is more fun```

Is displayed in the Discord web client as:

<pre><code>haikus written here
really shouldn't cause conflict
markdown is more fun</code></pre>

and displayed in libpurple clients as:

<pre><code>written here
really shouldn't cause conflict
markdown is more fun</code></pre>

Unlike other markdown variants, Discord does not treat the opening word
or line of a fenced code block as an [info
string](https://spec.commonmark.org/0.28/#info-string).

This commit adjusts the behavior to match the Discord web client (and
presumably the other Discord clients, although I have only tested the
web client) by displaying the first word after the start of a fenced
code block.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>