markdown: Sync with Sapphire's markdown.c
The markdown library internally developed with libdiscord is useful to
other Markdown-aware members of the libpurple family, including
Sapphire. Accordingly, it is split into its own dedicated file, to
facilitate merging with sidestream implementations in the future.

Sidestream changes fix a number of bugs in the existing Markdown
implementation, including:

	- Shruggies now work
	- Other backslash/etc abusing emoticons work

Other changes may not be of immediate use to purple-discord and are
accordingly disabled.