Disable join/part messages, closes #171
Join/part messages in guilds are incredibly annoying (especially for
large guilds where they cause flooding), and... serve essentially no
purpose to the user. On a stateful protocol like IRC, it makes sense to
track this. On something fluid like Discord where presence is a lot more
fuzzy, it's just... annoying, and something of a bandwidth hog perhaps.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>