Don't bundle icon pngs; generate with ImageMagick
Certain distributions, including Debian, have policies against including
binaries that are not in their preferred form of modification. For us,
that means the png logos are considered binary data, since we're
generating them from the svg. Therefore, require a dev dependency on
ImageMagick to convert from svg to png, and remove the pngs from the
repo.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>