Convert multi-line headers of flow control constructs to indented style
This makes it more obvious where the condition of an if statement ends,
more obvious where the control of a for loop ends by separating the
headers of each from the statements that they control.

This makes it more obvious where the arguments to a function end.

Also avoids overly deep indentation and spaces after tab indentation.